refactor: extract error handlers to neteye/error_handlers.py
